requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
bleach==6.1.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.1
cssselect2==0.7.0
defusedxml==0.8.0rc2
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-rest-swagger==2.2.0
django-taggit==5.0.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf_api_logger==1.1.15
fonttools==4.53.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.0
gunicorn==21.2.0
html5lib==1.1
httplib2==0.22.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.2
lxml==5.2.2
MarkupSafe==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
oscrypto==1.3.0
packaging==24.0
pdfkit==1.0.0
pillow==10.3.0
pipdeptree==2.13.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydyf==0.10.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyparsing==3.1.2
pypdf==4.2.0
pyphen==0.15.0
pypng==0.20220715.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
StrEnum==0.4.15
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.1.0
weasyprint==62.2
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3
